Add way to create posts symlink in dev
In development, this symlink will point to a clone of my
`personal-content--writings` repo.

When deploying, we will use a git subtree/submodule to access posts
instead. (That will come in a later commit.)

The `src/content/posts` directory is ignored from the repo so we can
make this switch.

Also, remove `.astro` from the repo as we don't need this there either
(it's autogenerated).