Add Gallery component
This commit adds a gallery component which can be used to show a set of
images within a blog post. Clicking on an image will show an enlarged
version of that image.

Note that the gallery is implemented with SolidJS and not React. In
fact, I've removed the React integration entirely. I realized in
implementing the gallery that React doesn't really fit well with the
Astro way of doing things: because client components are
rendered twice, once on the server and once on the client, a component
with a `useEffect` may not function correctly (for some reason React
says that the Rules of Hooks have been violated). Switching to SolidJS
works better because components only get run once (not each time a
component is rendered) and signals don't need to exist within a
component to function (i.e. there are no Rules of Hooks).

I also learned that contexts (either React or Solid) don't work in Astro.
This is because each component is rendered in its own Astro island. To
get around this, Astro recommends using its `nanostores` library. (I
mention this because I first implemented the gallery by adding Gallery
and GalleryImage components to keep the JSX in the MDX documents as
compatible with HTML as possible. I later decided to revert that, so I'm
no longer using contexts, but it's still interesting nonetheless.)