Support Obsidian-style callouts
Obsidian adds its own support for callouts. These differ from GitHub's
alerts, because more types are supported, and nested and collapsible
callouts are supported as well. The `rehype-callouts` package recognizes
Obsidian's callout syntax, so we add that to the list of Rehype plugins.
(In the future I'd love to use something like
`remark-github-admonitions-to-directives` to turn Obsidian's callout
syntax into the more standard admonition syntax that Docusaurus and
others support. But we're not there yet.)